+++ src/components/survey/SurveySingleItemView/ResponseComponent/InputTypes/ResponsiveSingleChoiceArray.tsx
@@ -360,7 +360,7 @@ const ResponsiveSingleChoiceArray: React.FC<ResponsiveSingleChoiceArrayProps> = 
                   <input
                     className="form-check-input cursor-pointer"
                     type="radio"
-                    name={item.key}
+                    name={'tableMode-' + item.key}
                     onChange={radioSelectionChanged(item.key)}
                     value={oi.key}
                     checked={isResponseSet(item.key, oi.key)}