+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.9.2 - 2022-02-11
+
+- Fix input name issue in responsive single choice array and bipolar likert.
+
 ## 1.9.1 - 2022-02-11
 
 ### Changed